requirements.txt
coverage==5.3
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
graphene-django==2.13.0
gunicorn==20.0.4
mixer==6.1.3
psycopg2-binary==2.8.6

PyJWT==1.7.1